third_party/blink/web_tests/platform/generic/fast/workers/worker-onerror-09-expected.txt

This tests that unhandled exceptions in postMessage() are delivered to 'worker.onerror'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Page-level worker.onerror handler triggered:
PASS errorEvent.message is "Uncaught DataCloneError: Failed to execute 'postMessage' on 'DedicatedWorkerGlobalScope': Value at index 0 does not have a transferable type."
PASS stripURL(errorEvent.filename) is "[blob: URL]"
PASS errorEvent.lineno is 3
PASS errorEvent.colno is 9
PASS errorEvent.error is null
PASS successfullyParsed is true

TEST COMPLETE
